requirements.txt
fastapi
uvicorn
pillow
transformers
torch
torchvision
ultralytics
python-multipart
